+++ backend/src/Faktur.Domain/Products/Events/ProductUpdatedEvent.cs
@@ -17,7 +17,7 @@ public record ProductUpdatedEvent : DomainEvent, INotification
 
   public Modification<FlagsUnit>? Flags { get; set; }
 
-  public Modification<double?>? UnitPrice { get; set; }
+  public Modification<decimal?>? UnitPrice { get; set; }
   public Modification<UnitType?>? UnitType { get; set; }
 
   public bool HasChanges => DepartmentNumber != null
